+++ app/src/main/java/com/stopsmoke/kekkek/presentation/MainActivity.kt
@@ -1,10 +1,11 @@
-package com.stopsmoke.kekkek
+package com.stopsmoke.kekkek.presentation
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.stopsmoke.kekkek.R
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
