+++ app/src/main/java/com/stopsmoke/kekkek/presentation/community/CommunityFragment.kt
@@ -1,4 +1,4 @@
-package com.stopsmoke.kekkek.community
+package com.stopsmoke.kekkek.presentation.community
 
 import android.os.Bundle
 import android.view.LayoutInflater
