+++ app/src/main/java/com/stopsmoke/kekkek/presentation/my/MyFragment.kt
@@ -1,4 +1,4 @@
-package com.stopsmoke.kekkek.myFragment
+package com.stopsmoke.kekkek.presentation.my
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
-import com.stopsmoke.kekkek.R
 import com.stopsmoke.kekkek.databinding.FragmentMyBinding
