+++ app/src/main/java/com/stopsmoke/kekkek/presentation/my/MyUiState.kt
@@ -1,4 +1,4 @@
-package com.stopsmoke.kekkek.myFragment
+package com.stopsmoke.kekkek.presentation.my
 
 data class MyUiState(
     val myLoginUiState: MyLoginStatusState
@@ -11,7 +11,7 @@ data class MyUiState(
 sealed interface MyLoginStatusState {
     data object NeedLoginUiState: MyLoginStatusState
 
-    sealed class LoggedUiState : MyLoginStatusState{ // 게스트 로그인 기능 추가 경우? 갑자기 생각
+    sealed class LoggedUiState : MyLoginStatusState { // 게스트 로그인 기능 추가 경우? 갑자기 생각
         data class MyIdLoggedUiState(val myItem: MyItem)
         //data class GuestIdLoggedUiState(val guestItem: MyItem)
     }
