--- app/src/main/java/com/stopsmoke/kekkek/snsLoginTest/SnsTestActivity.kt
@@ -1,110 +1,0 @@
-package com.stopsmoke.kekkek.snsLoginTest
-
-import android.os.Bundle
-import android.view.View
-import android.widget.Toast
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.navercorp.nid.NaverIdLoginSDK
-import com.navercorp.nid.oauth.NidOAuthLogin
-import com.navercorp.nid.oauth.OAuthLoginCallback
-import com.navercorp.nid.profile.NidProfileCallback
-import com.navercorp.nid.profile.data.NidProfileResponse
-import com.stopsmoke.kekkek.R
-import com.stopsmoke.kekkek.databinding.ActivitySnsTestBinding
-
-class SnsTestActivity : AppCompatActivity() {
-    private val binding:ActivitySnsTestBinding by lazy {
-        ActivitySnsTestBinding.inflate(layoutInflater)
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-
-        naverInit()
-    }
-
-    private fun naverInit(){
-        val naverClientId = getString(R.string.social_login_info_naver_client_id)
-        val naverClientSecret = getString(R.string.social_login_info_naver_client_secret)
-        val naverClientName = getString(R.string.social_login_info_naver_client_name)
-        NaverIdLoginSDK.initialize(this, naverClientId, naverClientSecret , naverClientName)
-
-        setLayoutState(false)
-
-        binding.tvNaverLogin.setOnClickListener {
-            startNaverLogin()
-        }
-        binding.tvNaverLogout.setOnClickListener {
-            startNaverLogout()
-        }
-        binding.tvNaverDeleteToken.setOnClickListener {
-            startNaverDeleteToken()
-        }
-    }
-
-    private fun startNaverLogin(){
-        var naverToken :String? = ""
-
-        val profileCallback = object : NidProfileCallback<NidProfileResponse> {
-            override fun onSuccess(response: NidProfileResponse) {
-                val userId = response.profile?.id
-                binding.tvResult.text = "id: ${userId} \ntoken: ${naverToken}"
-                setLayoutState(true)
-                Toast.makeText(this@SnsTestActivity, "네이버 아이디 로그인 성공!", Toast.LENGTH_SHORT).show()
-            }
-            override fun onFailure(httpStatus: Int, message: String) {
-                val errorCode = NaverIdLoginSDK.getLastErrorCode().code
-                val errorDescription = NaverIdLoginSDK.getLastErrorDescription()
-                Toast.makeText(this@SnsTestActivity, "errorCode: ${errorCode}\n" +
-                        "errorDescription: ${errorDescription}", Toast.LENGTH_SHORT).show()
-            }
-            override fun onError(errorCode: Int, message: String) {
-                onFailure(errorCode, message)
-            }
-        }
-
-        /** OAuthLoginCallback을 authenticate() 메서드 호출 시 파라미터로 전달하거나 NidOAuthLoginButton 객체에 등록하면 인증이 종료되는 것을 확인할 수 있습니다. */
-        val oauthLoginCallback = object : OAuthLoginCallback {
-            override fun onSuccess() {
-                // 네이버 로그인 인증이 성공했을 때 수행할 코드 추가
-                naverToken = NaverIdLoginSDK.getAccessToken()
-
-                //로그인 유저 정보 가져오기
-                NidOAuthLogin().callProfileApi(profileCallback)
-            }
-            override fun onFailure(httpStatus: Int, message: String) {
-                val errorCode = NaverIdLoginSDK.getLastErrorCode().code
-                val errorDescription = NaverIdLoginSDK.getLastErrorDescription()
-                Toast.makeText(this@SnsTestActivity, "errorCode: ${errorCode}\n" +
-                        "errorDescription: ${errorDescription}", Toast.LENGTH_SHORT).show()
-            }
-            override fun onError(errorCode: Int, message: String) {
-                onFailure(errorCode, message)
-            }
-        }
-
-        NaverIdLoginSDK.authenticate(this, oauthLoginCallback)
-    }
-
-    private fun startNaverLogout(){}
-
-    private fun startNaverDeleteToken(){
-    }
-
-    private fun setLayoutState(login: Boolean){
-        if(login){
-            binding.tvNaverLogin.visibility = View.GONE
-            binding.tvNaverLogout.visibility = View.VISIBLE
-            binding.tvNaverDeleteToken.visibility = View.VISIBLE
-        }else{
-            binding.tvNaverLogin.visibility = View.VISIBLE
-            binding.tvNaverLogout.visibility = View.GONE
-            binding.tvNaverDeleteToken.visibility = View.GONE
-            binding.tvResult.text = ""
-        }
-    }
-}
