+++ app/src/main/res/layout/fragment_community.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="20dp"
-    tools:context=".community.CommunityFragment">
+    tools:context=".presentation.community.CommunityFragment">
 
     <ImageView
         android:id="@+id/iv_community_bell"
